requirements.txt
Pillow==10.2.0
slack_sdk==3.26.2
